Extracted find_amber_files function to utils to reuse
